+++ app/GraphQL/Inventory/Builders/Products/ProductBuilder.php
@@ -8,11 +8,10 @@ use GraphQL\Type\Definition\ResolveInfo;
 use Illuminate\Database\Eloquent\Builder;
 use Kanvas\Apps\Models\Apps;
 use Kanvas\Companies\Models\CompaniesBranches;
-use Kanvas\Connectors\RainForest\Workflows\SearchWorkflow as RainForestSearchWorkflow;
 use Kanvas\Inventory\Products\Models\Products;
 use Kanvas\Inventory\Regions\Models\Regions;
+use Kanvas\Workflow\Enums\WorkflowEnum;
 use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
-use Workflow\WorkflowStub;
 
 class ProductBuilder
 {
@@ -24,14 +23,19 @@ class ProductBuilder
     ): Builder {
         $app = app(Apps::class);
         $companyBranch = app(CompaniesBranches::class);
-        if ($companyBranch && key_exists('search', $args)) {
-            $region = Regions::getDefault($companyBranch->company);
-            $workflow = WorkflowStub::make(RainForestSearchWorkflow::class);
-            $workflow->start($app, auth()->user(), $companyBranch, $region, $args['search']);
-        }
-
         $company = auth()->user()->getCurrentCompany();
         $user = auth()->user();
+
+        if ($companyBranch && key_exists('search', $args)) {
+            $region = Regions::getDefault($companyBranch->company);
+            $app->fireWorkflow(event: WorkflowEnum::SEARCH->value, params: [
+                'app' => $app,
+                'user' => auth()->user(),
+                'companyBranch' => $companyBranch,
+                'region' => $region,
+                'search' => $args['search'],
+            ]);
+        }
 
         if (! $user->isAppOwner()) {
             //Products::setSearchIndex($company->getId());
