+++ src/Domains/Connectors/RainForest/Enums/ConfigurationEnum.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Kanvas\Connectors\RainForest\Enums;
 
-enum RainForestEnum: string
+enum ConfigurationEnum: string
 {
     case AMAZON_ID = 'AMAZON_ID';
     case WEIGHT_UNIT = 'GRAMS';
