+++ src/Domains/Connectors/Shopify/Services/ShopifyInventoryService.php
@@ -6,7 +6,7 @@ namespace Kanvas\Connectors\Shopify\Services;
 
 use Baka\Contracts\AppInterface;
 use Baka\Contracts\CompanyInterface;
-use Kanvas\Connectors\RainForest\Enums\RainForestEnum;
+use Kanvas\Connectors\RainForest\Enums\ConfigurationEnum;
 use Kanvas\Connectors\Shopify\Client;
 use Kanvas\Connectors\Shopify\Enums\StatusEnum;
 use Kanvas\Inventory\Channels\Models\Channels;
@@ -61,7 +61,7 @@ class ShopifyInventoryService
                 if ($variant->getShopifyId($this->warehouses->regions) === null) {
                     $variant->setShopifyId($this->warehouses->regions, $shopifyVariant['id']);
                 }
-                $this->setStock($variant, $channel);
+                // $this->setStock($variant, $channel);
             }
         } else {
             $shopifyProduct = $this->shopifySdk->Product($shopifyProductId);
@@ -69,7 +69,7 @@ class ShopifyInventoryService
 
             foreach ($product->variants as $variant) {
                 $this->saveVariant($variant, $channel);
-                $this->setStock($variant, $channel);
+                // $this->setStock($variant, $channel);
             }
         }
 
@@ -118,11 +118,11 @@ class ShopifyInventoryService
             'compare_at_price' => $discountedPrice ?? 0,
             'inventory_policy' => 'deny',
             'published' => $price > 0,
-            'grams' => $variant->get(RainForestEnum::WEIGHT_UNIT->value) ?? 453.592,
+            'grams' => $variant->get(ConfigurationEnum::WEIGHT_UNIT->value) ?? 453.592,
         ];
 
         if ($quantity > 0) {
-            // $this->setStock($variant, $channel);
+            $this->setStock($variant, $channel);
         }
 
         if ($variant->product->getShopifyId($this->warehouses->regions)) {
