+++ src/Domains/Inventory/Variants/Enums/ConfigurationEnum.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Kanvas\Inventory\Variants\Enums;
+
+enum ConfigurationEnum: string
+{
+    case WEIGHT_UNIT = 'GRAMS';
+}
